Add files via upload
This is the overall skeleton of the experiment - my part and Alex's part. I haven't incorporated button presses yet. But I think it's important to have the outline in the current version saved as it has all the important settings for the sound and screen. Please check it out if it runs properly on your local console. If there are any bugs please report them.