Add files via upload
Creating_aud_stimuli is a function which takes in a vector of frequencies and two scalars (the first indicating the desired difference between the 1st and the 2nd frequency and the second indicating the desired difference between the 2nd and the 3rd frequency). The function returns a matrix of all possible combinations following these rules. 

Stimuli is an example of the selected triplets (non-randomised), which were adopted in our experiment.